+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+### Unreleased
+
+ * Added `Account[Create|Update]Transaction.[get|set]maxAutomaticTokenAssociations`
+ * Added `TokenAssociation` and `TransactionRecord.automaticTokenAssociations`
+ * Added `AccountInfo.maxAutomaticTokenAssociations`
+ * Added `TokenRelationship.automaticAssociation`
+
 ### v2.0.12
 
 ### Added